1. fix bug, https://github.com/zhangzhuang15/JasonPackages/issues/6 2. refactor code and make it more extensive 3. add settings configuration for future extending
