+++ packages/vscode-plugin-dictionary/src/translation-impl/translator/index.ts
@@ -1,0 +1,12 @@
+import { 
+    name as icibaName, 
+    requestContextBuilder as icibaRequestContextBuilder, 
+    translate as icibaTranslate
+} from "@/translation-impl/translator/iciba/index";
+
+export default {
+    [icibaName]: {
+        requestContextBuilder: icibaRequestContextBuilder, 
+        translate: icibaTranslate 
+    }
+}
