+++ packages/vscode-plugin-dictionary/src/translation-impl/utils/httpclient.util.ts
@@ -1,0 +1,3 @@
+export { request as httpClient } from "node:http";
+export { request as httpsClient } from "node:https";
+export { connect as http2Client, OutgoingHttpHeaders as Http2Headers} from "node:http2";
